♻️ refactor(test): reorganize test fakes to tests/fakes directory
- Move FakeUnitOfWork to tests/fakes for better organization
- Create FakeQuizRepository to satisfy AppDependencies interface
- Update all imports to use the new location
- Add missing methods to FakeUserRepository for consistency

This centralizes all test fakes in one location, making them easier
to find and maintain.
