+++ apps/api/test-utils/db/index.ts
@@ -2,16 +2,16 @@
  * Database utilities barrel export
  */
 
-export { createTestDb, withTestDb } from './connection';
-export { closeAllTrackedClients } from './container';
-export { createTestDatabase, type TestDatabaseOptions } from './core';
 export {
   FakeQuizRepository,
   FakeUnitOfWork,
   FakeUnitOfWorkFactory,
   FakeUserRepository,
   withFakeUnitOfWork,
-} from './FakeUnitOfWork';
+} from '../../tests/fakes';
+export { createTestDb, withTestDb } from './connection';
+export { closeAllTrackedClients } from './container';
+export { createTestDatabase, type TestDatabaseOptions } from './core';
 export { drizzleMigrate, resetMigrationState, verifyMigrationTables } from './migrations';
 export { testSchema } from './schema';
 export { seedAdminUser, seedUsers } from './seeds';
