+++ apps/api/tests/e2e/auth-login.test.ts
@@ -10,17 +10,19 @@ import { UserRole } from '@api/features/auth/domain/value-objects/UserRole';
 import { StubAuthProvider } from '@api/infra/auth/StubAuthProvider';
 import { unwrapOrFail } from '@api/test-support';
 import { beforeEach, describe, expect, it } from 'vitest';
-import { FakeUserRepository } from '../fakes/FakeUserRepository';
+import { FakeQuizRepository, FakeUserRepository } from '../fakes';
 import { fakeLogger } from '../helpers/app';
 
 describe('POST /api/auth/login - E2E', () => {
   let app: ReturnType<typeof buildApp>;
   let fakeUserRepo: FakeUserRepository;
+  let fakeQuizRepo: FakeQuizRepository;
   let stubAuthProvider: StubAuthProvider;
 
   beforeEach(() => {
     // Create fresh fake dependencies for each test
     fakeUserRepo = new FakeUserRepository();
+    fakeQuizRepo = new FakeQuizRepository();
     stubAuthProvider = new StubAuthProvider();
 
     // Build app with all required dependencies
@@ -31,6 +33,7 @@ describe('POST /api/auth/login - E2E', () => {
         // No-op for tests
       },
       userRepository: fakeUserRepo,
+      quizRepository: fakeQuizRepo,
       authProvider: stubAuthProvider,
     });
   });
