+++ apps/api/tests/fakes/FakeUserRepository.ts
@@ -50,17 +50,18 @@ export class FakeUserRepository implements IUserRepository {
     this.users.set(user.email.toString(), user);
   }
 
-  async isEmailTaken(email: Email): Promise<boolean> {
-    return this.users.has(email.toString());
+  async isEmailTaken(email: Email, excludeUserId?: UserId): Promise<boolean> {
+    const user = await this.findByEmail(email);
+    if (!user) return false;
+    if (excludeUserId && user.id === excludeUserId) return false;
+    return true;
   }
 
-  async isUsernameTaken(username: string): Promise<boolean> {
-    for (const user of this.users.values()) {
-      if (user.username === username) {
-        return true;
-      }
-    }
-    return false;
+  async isUsernameTaken(username: string, excludeUserId?: UserId): Promise<boolean> {
+    const user = await this.findByUsername(username);
+    if (!user) return false;
+    if (excludeUserId && user.id === excludeUserId) return false;
+    return true;
   }
 
   // Test helper methods
@@ -84,6 +85,13 @@ export class FakeUserRepository implements IUserRepository {
    */
   getAllUsers(): User[] {
     return Array.from(this.users.values());
+  }
+
+  /**
+   * Get all users (test helper) - alias for getAllUsers
+   */
+  getAll(): User[] {
+    return this.getAllUsers();
   }
 
   /**
