+++ apps/api/tests/fakes/index.ts
@@ -3,8 +3,13 @@
  * @fileoverview Centralized exports for all fake implementations
  */
 
+export { FakeQuizRepository } from './FakeQuizRepository';
+export {
+  FakeUnitOfWork,
+  FakeUnitOfWorkFactory,
+  withFakeUnitOfWork,
+} from './FakeUnitOfWork';
 export { FakeUserRepository } from './FakeUserRepository';
 
 // TODO: Add more fake implementations as needed
-// export { FakeQuizRepository } from './FakeQuizRepository';
 // export { FakeEmailService } from './FakeEmailService';
